+++ Model/ShopeeProductUpdateStockModel.cs
@@ -8,7 +8,7 @@ namespace WeTools.ShopeeSDK.Model
     public class ShopeeProductUpdateStockFailureModel
     {
         [JsonProperty("model_id")]
-        public int ModelId { get; set; }
+        public long ModelId { get; set; }
 
         [JsonProperty("failed_reason")]
         public string FailedReason { get; set; }
@@ -17,7 +17,7 @@ namespace WeTools.ShopeeSDK.Model
     public class ShopeeProductUpdateStockSuccessModel
     {
         [JsonProperty("model_id")]
-        public int ModelId { get; set; }
+        public long ModelId { get; set; }
 
         [JsonProperty("original_price")]
         public double OriginalPrice { get; set; }
